server: split off ldap/search from ldap_handler
